requirements.txt
flask
flask_limiter
flask_restful
pixivpy
requests
sqlalchemy
mysqlclient